+++ superwall/src/main/java/com/superwall/sdk/paywall/vc/PaywallView.kt
@@ -55,7 +55,7 @@ import com.superwall.sdk.models.triggers.TriggerRuleOccurrence
 import com.superwall.sdk.network.device.DeviceHelper
 import com.superwall.sdk.paywall.manager.PaywallCacheLogic
 import com.superwall.sdk.paywall.manager.PaywallManager
-import com.superwall.sdk.paywall.manager.PaywallViewControllerCache
+import com.superwall.sdk.paywall.manager.PaywallViewCache
 import com.superwall.sdk.paywall.presentation.PaywallCloseReason
 import com.superwall.sdk.paywall.presentation.PaywallInfo
 import com.superwall.sdk.paywall.presentation.get_presentation_result.internallyGetPresentationResult
@@ -67,8 +67,8 @@ import com.superwall.sdk.paywall.presentation.result.PresentationResult
 import com.superwall.sdk.paywall.vc.Survey.SurveyManager
 import com.superwall.sdk.paywall.vc.Survey.SurveyPresentationResult
 import com.superwall.sdk.paywall.vc.delegate.PaywallLoadingState
-import com.superwall.sdk.paywall.vc.delegate.PaywallViewControllerDelegateAdapter
-import com.superwall.sdk.paywall.vc.delegate.PaywallViewControllerEventDelegate
+import com.superwall.sdk.paywall.vc.delegate.PaywallViewDelegateAdapter
+import com.superwall.sdk.paywall.vc.delegate.PaywallViewEventDelegate
 import com.superwall.sdk.paywall.vc.web_view.PaywallMessage
 import com.superwall.sdk.paywall.vc.web_view.SWWebView
 import com.superwall.sdk.paywall.vc.web_view.SWWebViewDelegate
@@ -86,18 +86,18 @@ import java.util.*
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 
-class PaywallViewController(
+class PaywallView(
     context: Context,
     override var paywall: Paywall,
-    val eventDelegate: PaywallViewControllerEventDelegate? = null,
-    var delegate: PaywallViewControllerDelegateAdapter? = null,
+    val eventDelegate: PaywallViewEventDelegate? = null,
+    var delegate: PaywallViewDelegateAdapter? = null,
     val deviceHelper: DeviceHelper,
     val factory: Factory,
     val storage: Storage,
     val paywallManager: PaywallManager,
     override val webView: SWWebView,
-    val cache: PaywallViewControllerCache?,
-    private val loadingViewController: LoadingViewController = LoadingViewController(context),
+    val cache: PaywallViewCache?,
+    private val loadingView: LoadingView = LoadingView(context),
 ) : FrameLayout(context),
     PaywallMessageHandlerDelegate,
     SWWebViewDelegate,
@@ -222,7 +222,7 @@ class PaywallViewController(
         hideShimmerView()
 
         // Add the loading view and hide it
-        addView(loadingViewController)
+        addView(loadingView)
         hideLoadingView()
 
         setBackgroundColor(backgroundColor)
@@ -425,7 +425,7 @@ class PaywallViewController(
             paywallResult = result,
             paywallCloseReason = closeReason,
             activity = encapsulatingActivity,
-            paywallViewController = this,
+            paywallView = this,
             loadingState = loadingState,
             isDebuggerLaunched = request?.flags?.isDebuggerLaunched == true,
             paywallInfo = info,
@@ -518,7 +518,7 @@ class PaywallViewController(
 
     override fun eventDidOccur(paywallEvent: PaywallWebEvent) {
         CoroutineScope(Dispatchers.IO).launch {
-            eventDelegate?.eventDidOccur(paywallEvent, this@PaywallViewController)
+            eventDelegate?.eventDidOccur(paywallEvent, this@PaywallView)
         }
     }
 
@@ -551,13 +551,13 @@ class PaywallViewController(
             return
         }
         CoroutineScope(Dispatchers.Main).launch {
-            loadingViewController.visibility = View.VISIBLE
+            loadingView.visibility = View.VISIBLE
         }
     }
 
     private fun hideLoadingView() {
         CoroutineScope(Dispatchers.Main).launch {
-            loadingViewController.visibility = View.GONE
+            loadingView.visibility = View.GONE
         }
     }
 
@@ -690,7 +690,7 @@ class PaywallViewController(
             val trackedEvent =
                 InternalSuperwallEvent.PaywallWebviewLoad(
                     state = InternalSuperwallEvent.PaywallWebviewLoad.State.Start(),
-                    paywallInfo = this@PaywallViewController.info,
+                    paywallInfo = this@PaywallView.info,
                 )
             Superwall.instance.track(trackedEvent)
         }
@@ -720,13 +720,13 @@ class PaywallViewController(
         } catch (e: MalformedURLException) {
             Logger.debug(
                 logLevel = LogLevel.debug,
-                scope = LogScope.paywallViewController,
+                scope = LogScope.paywallView,
                 message = "Invalid URL provided for \"Open In-App URL\" click behavior.",
             )
         } catch (e: Throwable) {
             Logger.debug(
                 logLevel = LogLevel.debug,
-                scope = LogScope.paywallViewController,
+                scope = LogScope.paywallView,
                 message = "Exception thrown for \"Open In-App URL\" click behavior.",
             )
         }
@@ -741,13 +741,13 @@ class PaywallViewController(
         } catch (e: MalformedURLException) {
             Logger.debug(
                 logLevel = LogLevel.debug,
-                scope = LogScope.paywallViewController,
+                scope = LogScope.paywallView,
                 message = "Invalid URL provided for \"Open External URL\" click behavior.",
             )
         } catch (e: Throwable) {
             Logger.debug(
                 logLevel = LogLevel.debug,
-                scope = LogScope.paywallViewController,
+                scope = LogScope.paywallView,
                 message = "Exception thrown for \"Open External URL\" click behavior.",
             )
         }
@@ -771,7 +771,7 @@ class PaywallViewController(
         webView.evaluateJavascript("window.paywall.accept([$payload])", null)
         Logger.debug(
             logLevel = LogLevel.debug,
-            scope = LogScope.paywallViewController,
+            scope = LogScope.paywallView,
             message = "Game controller event occurred: $payload",
         )
     }
@@ -809,7 +809,7 @@ class SuperwallPaywallActivity : AppCompatActivity() {
 
         fun startWithView(
             context: Context,
-            view: PaywallViewController,
+            view: PaywallView,
             presentationStyleOverride: PaywallPresentationStyle? = null,
         ) {
             val key = UUID.randomUUID().toString()
@@ -868,7 +868,7 @@ class SuperwallPaywallActivity : AppCompatActivity() {
         }
 
         val view =
-            ViewStorage.retrieveView(key) as? PaywallViewController ?: run {
+            ViewStorage.retrieveView(key) as? PaywallView ?: run {
                 finish() // Close the activity if the view associated with the key is not found
                 return
             }
@@ -919,7 +919,7 @@ class SuperwallPaywallActivity : AppCompatActivity() {
 
     override fun onStart() {
         super.onStart()
-        val paywallVc = contentView as? PaywallViewController ?: return
+        val paywallVc = contentView as? PaywallView ?: return
 
         if (paywallVc.isSafariVCPresented) {
             paywallVc.isSafariVCPresented = false
@@ -930,7 +930,7 @@ class SuperwallPaywallActivity : AppCompatActivity() {
 
     override fun onResume() {
         super.onResume()
-        val paywallVc = contentView as? PaywallViewController ?: return
+        val paywallVc = contentView as? PaywallView ?: return
 
         paywallVc.viewDidAppear()
     }
@@ -938,7 +938,7 @@ class SuperwallPaywallActivity : AppCompatActivity() {
     override fun onPause() {
         super.onPause()
 
-        val paywallVc = contentView as? PaywallViewController ?: return
+        val paywallVc = contentView as? PaywallView ?: return
 
         CoroutineScope(Dispatchers.Main).launch {
             paywallVc.viewWillDisappear()
@@ -948,7 +948,7 @@ class SuperwallPaywallActivity : AppCompatActivity() {
     override fun onStop() {
         super.onStop()
 
-        val paywallVc = contentView as? PaywallViewController ?: return
+        val paywallVc = contentView as? PaywallView ?: return
 
         CoroutineScope(Dispatchers.Main).launch {
             paywallVc.viewDidDisappear()