+++ superwall/src/main/java/com/superwall/sdk/store/StoreKitManagerInterface.kt
@@ -6,7 +6,7 @@ import com.superwall.sdk.models.paywall.PaywallProducts
 import com.superwall.sdk.models.product.Product
 import com.superwall.sdk.models.product.ProductItem
 import com.superwall.sdk.models.product.ProductVariable
-import com.superwall.sdk.paywall.vc.PaywallViewController
+import com.superwall.sdk.paywall.vc.PaywallView
 import com.superwall.sdk.store.abstractions.product.StoreProduct
 
 data class GetProductsResponse(
@@ -27,11 +27,11 @@ interface StoreKitManagerInterface {
         substituteProducts: PaywallProducts? = null,
     ): GetProductsResponse
 
-    suspend fun tryToRestore(paywallViewController: PaywallViewController)
+    suspend fun tryToRestore(paywallView: PaywallView)
 
     suspend fun processRestoration(
         restorationResult: RestorationResult,
-        paywallViewController: PaywallViewController,
+        paywallView: PaywallView,
     )
 
     suspend fun refreshReceipt()